Link skill requirements to the skill
Most of the skill reqs are named wrong, so almost none of them actually
link. We'll go through the rulebook and fix those later.

I change 'clazz' to 'class' in json, which has to be corrected in the
API scripts too. Also just use 'class' for naming when possible.

Also put the colon back in skill names, and update API scripts
accordingly.