Clean up the mess that was RollTime
Some effects modify stats and damages. Previously, rolls for these two
concepts were kept separate by differing RollTimes. There is no good
reason for that. Really, RollTime exists for effects that need to know
whether or not a roll was a crit. We don't need to separate out effects
that modify stats, skills, concentration, initiative, and damage.

There are now two RollTimes - one for before we know if a roll is a crit
and one for after. The vast majority of effects can be applied before we
know if a roll is a crit, since they don't care about it. Only the few
that do care should have the latter RollTime. This lets us compress most
rolls into RollTime.DEFAULT and still get all the info we need.

This cleans up stat rolls, skill rolls, and concentration rolls. It lets
us fix abilities that create effects that should modify stats and
damages - now it's just one effect for both rather than two separate
ones.

Side fixes:
* Implement initiative bonus and rolls
* Fix a todo about stat multipliers, those have been added
* Fix up the hackiness around determining crit chance from items