Support spells that have crit change
Rolls have a new internal field "should_apply_crit". This is true for
physical rolls, and false for other types. Persistent effects can set
this to true though, like for an ability that allows magic-type attacks
to crit. Implement Magic Primer, which does this.

All spells have to roll crit chance now, as a result of this. Most of
the time that roll and related effects won't get applied because of the
new field in rolls. Magic Dart becomes a bit more complicated as a
result.

Stat effects have to specify an applicable roll type now so that things
like a dagger's crit chance won't get applied to a magic attack.