Make Power Spike work for other parts of rolls
Power Spike now also works for effects that:
* Improve stats
* Improve skills
* Increase hidden stat bonuses
* Increase crit change
* Increase crit damage mod
* Increase initiative bonus
* Increase concentration bonus

It won't work for something like Assassin::Sharpen, which changes d4s to
d6s. I don't even know what Power Spiking that buff would do - change to
d7s? In any case, it wouldn't be recognized because the fake roll used
for Power Spike doesn't have any dice to begin with, so Sharpen's effect
wouldn't modify the roll.

All of this is done in the hope that this madness for Power Spike can be
reused for generic buff efficiency.