Better buff efficiency handling
Revise an efficient buff when it is applied, rather than when it is
created. Store an 'efficiency' parameter with each persistent effect to
support this. This makes modifying efficiency much easier, like for
Power Spike, and also passed on the proper parameters (those used when
an attack gets the buff applied) when modifying an effect.