Made Alter Course more officially handled
Alter Course is unpleasant because it requires us to get the result of a
roll and then apply a multiplier to it. And it can be applied to any
ability, so it must be an arbitrary parameter.

This adds first-class support for arbitrary parameters like this. Now
they "takeover" the completion of an ability. The normal do_roll() path
will stop, assuming the arbitrary parameter handler will do whatever is
necessary to finish up. Alter Course itself does a fake roll to figure
out the real values of damages, then copies the results into new rolls
per redirection.

Also clean up other abilities and documentation added recently.