Put the calendar in the rulebook
Dynamically generate a table for the year calendar based on what day of
the week the first day of the years on. Also serve a page per month,
with holidays filled in based on the new rulebook file. A bunch of
styles had to change to make this look decent.

Also rename index.html template to rulebook.html, since that's what it
actually is.