Support fully specifying items in the character sheet
Maintaining a hardcoded list of items is an awful pattern. The number of
items will only grow, and we'd have to add in every single one to allow
it to work in the API. Now, if an item starts with "construct", we will
try to build an item off of the rest of the parameters specified in the
line. This lets the players fill out the effects of their own items, and
the API will figure out what the item is supposed to do on the fly,
rather than referring to a hardcoded list.

Only transitioned my own items for the moment. I will eventually go
through others and set them up like this as well. And then I'll remove
the "construct" keyword, make this the default, and document this new
strategy on the API webpage.

Side fixes.
* Fixed the fact that crit chance was duplicated in two places in a
  roll, leading to some crit chance not being included in pure stat
  rolls. Ditch the separate Roll.crit_chance class member in favor
  of the stat array.
* Allow a separate source to be specified for Alter Course, in case it
  is redirecting another player's attack.
* Correct rolling items directly, and support the new format for items.
* Use helper for roll-type checks in effects.