Support buff/enchant effectiveness
Add fields to a rolls for buff and enchant effectiveness. Make new
Effect handlers to add to these two. Fix todos in existing items for
buff/enchant effectiveness. Support constructing items with buff/enchant
effectiveness.

When we add a persistent effect, we'll use a fake roll to figure out the
caster's buff/enchant effectiveness. Check tags on an ability to see if
we should add buff or enchant effectiveness to the persistent effect.