Update stats and skills based on constructed items
Since we are defining items and their affixes on the character sheet now, we
can actually also update stats and skills automatically as items change. Item
affixes that boost stats are now reflected in the character sheet stats
section. Item affixes that boost skills are not reflected in the character
sheet skills section. Fortunately, we never poll the stat value or skill
rank/bonus directly in the API (we instead rely on attributes and AP in
skills), so this change should have no impact on how the API works.