Apply weapons shields rank 1 effect automatically
The Weapon Mastery: Shield rank 1 bonus is a 50% physical damage increase on
all attacks with shields. We would like this (and similar things in the future)
to be applied to rolls automatically, so that people can be brain dead. When
adding items to a roll, we'll also check what weapons are being used and if
the character has a skill at a rank we care about. If so, then adjust the roll
accordingly.

This also does some sick caching so that we only have to fetch attributes via
the Roll20 API once, so hopefully this isn't actually do costly to do
regularly for abilities.

Also update classes in the API.