Add a button to copy abilities from the rulebook
Players often copy abilities from the rulebook and dump it onto their Roll20
character sheet. Add a hidden table for each ability that has nice formatting
and a button to select and save the table to the clipboard. We need a separate
table because, if you copy the regular one, you get white text and borders,
which doesn't look right on the Roll20 character sheet. The hidden table has
black text, black borders, a better font, and a set width so that it looks
nice when pasted on Roll20.

Misc:
* Remove sub-coins from the character sheet, there's only GP.
* Check that character isn't undefined before we try creating one. This was
  showing up when changing turn order for some reason.
* Remove empty tags on abilities