Fix null items when checking for skill bonuses
When we do the fake roll to calculate buff/enchant effectiveness, we use a
roll with type ALL. This fell into the "else" case when checking for skill
bonuses, so we were checking both main hand and offhand weapons. This would
be ok, but we need to check that what we're adding isn't null so we don't
break things down the chain. And more broadly, we shouldn't be adding items
to be applied if they are null.