Show ki gained and healing from Ki Monk abilities
Spirit Punch is supposed to gain Ki equal to the 4dU physical damage roll. We
can do this roll secretly, add the result to the actual roll, and then also
show the Ki gained using the same result. We can do the same kind of thing
for Drain Punch as well.