Print description of abilities that the API doesn't know how to handle
Lookup the ability in the master class list first. If it's not found, the
player either typed it wrong or we need to update the list (manual process).
If we find the ability, see if we have a processor for it. If not, chat a
warning and just print the ability. If we do have a processor, run it.