Support item affix that boosts all skills
It was already there, but there has to be a colon in the skill name
for it to be counted as a skill. Also document it. And document other
newer things.

Show progress for uploading files to the server and stop hanging the
GUI by running work in a thread.