Tie API commands to character sheets
Pass in the character name with all API commands. The character name is an
attribute on character sheets, so we pull it from the attribute when making
each API call. The character name is wrapped in $dollar signs$ to separate
it from the rest of the message.

In the API, this ties the "Character" we create for an API request to a
character sheet, rather than who is speaking the request in chat. For players,
this doesn't really have an effect, since players always speak as themselves
in chat. For the GM, however, this makes is much easier to run commands for
a character. In effect, clicking a button on a character's sheet will always
run the API for that character.