Revise the Take Damage roll
Attach the type to each given value so that it doesn't really matter
if we are missing some parts. Hopefully this'll stop Nevil from
screwing it up anymore by skipping through the damage types in under
a nanosecond.

Also ignore empty elements in character lists, so trailing commas won't
fail the whole ability.