Plumbing ability basically everywhere
Once upon a time, in a galaxy of circuitry far, far away, I wanted to
identify redirected abilities with this actual name and a "redirected N"
suffix. This meant that we needed to get the ability down to takeover
parameter handlers, which meant we needed to pass it to all arbitrary
parameter handlers, and to all crit roll calls. That's what this
change is.

I also fixed a couple abilities that were apparently missed when
I changed do_roll() to take the ability instead of just a name.
Fortunately, nobody ever used these abilities.